Moved Error classes out of api package
Verified that they weren't in this package in the 1.x connector. Just trying to minimize the number of things in the "api" package.
