+++ src/main/java/com/marklogic/mule/extension/ErrorType.java
@@ -1,10 +1,9 @@
-package com.marklogic.mule.extension.api;
+package com.marklogic.mule.extension;
 
 import org.mule.runtime.extension.api.error.ErrorTypeDefinition;
 
 public enum ErrorType implements ErrorTypeDefinition<ErrorType> {
 
-    // TODO Test that this works.
     CONNECTION_ERROR,
 
     XML_TRANSFORMER_ERROR;
