+++ src/main/java/com/marklogic/mule/extension/api/DocumentAttributes.java
@@ -21,6 +21,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.marklogic.client.io.DocumentMetadataHandle;
+import com.marklogic.mule.extension.ErrorType;
 import org.mule.runtime.extension.api.exception.ModuleException;
 import org.w3c.dom.Node;
